requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
requests==2.32.4
readability-lxml==0.8.4.1
python-multipart==0.0.20
beautifulsoup4==4.13.4
lxml==6.0.0
python-dotenv==1.1.1
lxml-html-clean==0.4.2
